Added the newest build and tested
It seems to be working fine however if there are any issues there is a list of files that could not be copied over. It seems these are all temporary folders and were just recreated when the application was launched but just in case I have created a word document with the names of them.