Added the final additive at the end of the day
It takes all the overall energy and then takes away the deduction value basically covering for wasted energy and it will succeed you if you are under the maximum energy and fail you if you are under the minimum. I also changed the end of the day to 10 for now to test it is working.